requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
speechrecognition==3.10.0
pyaudio==0.2.11
opencv-python==4.8.1.78
pytesseract==0.3.10
transformers==4.35.0
torch==2.1.0
spacy==3.7.2
mediapipe==0.10.7
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
